% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/g3_fit_combine.R
\name{bind_fit}
\alias{bind_fit}
\title{Combine multiple gadget.fit objects}
\usage{
bind_fit(fit_list)
}
\arguments{
\item{fit_list}{A list of multiple gadget.fit objects.}
}
\value{
A gadget.fit object
}
\description{
This function combines multiple gadget.fit objects into a single fit object.
An additional column 'id' is added to each element of the resulting list to identify
the individual gadget.fits
}
\details{
The 'id' column will be taken from names(fit_list). If fit_list has no names, the 'id'
column will be taken from the positions, e.g.,
fit_list = list(a = fit1, b = fit2) will be produces id's of 'a' and 'b', whereas
fit_list = list(fit1, fit2) will produce id's of '1' and '2'
}
