% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/g3_fit_combine.R
\name{bind_fit_components}
\alias{bind_fit_components}
\title{Combine individual elements from multiple gadget.fit objects}
\usage{
bind_fit_components(fit_list, component)
}
\arguments{
\item{fit_list}{A list of multiple gadget.fit objects.}

\item{component}{A string specifying which list element to extract (ie, see names(fit))}
}
\value{
A data frame that contains an 'id' column
}
\description{
This function extracts a specific element from multiple gadget.fit objects
and combines them into a single dataframe. The output has an additional column 'id'
that distinguishes each input gadget.fit object
}
\details{
The 'id' column will be taken from names(fit_list). If fit_list has no names, the 'id'
column will be taken from the positions, e.g.,
fit_list = list(a = fit1, b = fit2) will be produces id's of 'a' and 'b', whereas
fit_list = list(fit1, fit2) will produce id's of '1' and '2'
}
