% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/g3_optim.R
\name{check_params_out}
\alias{check_params_out}
\title{Internal function to check whether g3_optim has succeeded or crashed. Used by the gadgetutils g3_* functions.}
\usage{
check_params_out(param_list, id_name = NULL)
}
\arguments{
\item{param_list}{List of g3 parameters}

\item{id_name}{replacement name for 'id' in the output}
}
\value{
A data.frame that has collated the 'summary' attribute of each element of param_list.
}
\description{
\code{check_params_out} checks for NULLs and try-errors in a list of g3 parameters
}
