% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/g3_iterative_shortcut.R
\name{estimate_weights_adist}
\alias{estimate_weights_adist}
\title{Residual variance for abundance distributions}
\usage{
estimate_weights_adist(dat)
}
\arguments{
\item{dat}{Abundance distribution model report (as a dataframe)}
}
\description{
This function fits a loess model to abundance data and returns the residual variance of the model fit
}
