% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/g3_iterative_shortcut.R
\name{estimate_weights_cdist}
\alias{estimate_weights_cdist}
\title{Residual variance for catch distributions}
\usage{
estimate_weights_cdist(dat)
}
\arguments{
\item{dat}{Catch distribution model report (as a dataframe)}
}
\description{
Calculates the residual variance for catch distributions using a GLM
}
