% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/parameter_template_utils.R
\name{eval_bounded}
\alias{eval_bounded}
\title{Convert normalised bounded form back to value}
\usage{
eval_bounded(x, lower, upper)
}
\arguments{
\item{x}{Input normalised value}

\item{lower}{Lower bound for x}

\item{upper}{Upper bound for x}
}
\value{
Raw value for x without bounding
}
\description{
Convert normalised bounded form back to value
}
