% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/g3_fit.R
\name{extract_year_step}
\alias{extract_year_step}
\title{Extracts year and step columns}
\usage{
extract_year_step(data)
}
\arguments{
\item{data}{dataframe of gadget3 reports. Must include a 'time' column.}
}
\description{
Gadget3 reports have a time column in the format 'YEAR-STEP'. This function creates a 'year' and 'step' column by splitting the existing 'time' column. This is useful for processing gagdet3 reports.
}
