% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/action_utils.R
\name{g3_action_order}
\alias{g3_action_order}
\title{Given an action type, return it's running order in a g3 model}
\usage{
g3_action_order(action_name)
}
\arguments{
\item{action_name}{String for a action type, e.g. "natural_mortality"}
}
\value{
Position integer, e.g. 4
}
\description{
Given an action type, return it's running order in a g3 model
}
