% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/parameter_template_utils.R
\name{g3_add_parscale}
\alias{g3_add_parscale}
\title{Added parscale to a parameter template. No longer needed.}
\usage{
g3_add_parscale(parameters)
}
\arguments{
\item{parameters}{A gadget3 parameter dataframe}
}
\description{
Added parscale to a parameter template. No longer needed.
}
