% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/g3_fit.R
\name{g3_fit}
\alias{g3_fit}
\title{Build output summary suitable for feeding into gadgetplots}
\usage{
g3_fit(model, params, rec.steps = 1, steps = 1, adreport_re = "^$")
}
\arguments{
\item{model}{A G3 model built using \code{\link[gadget3]{g3_to_r}} or \code{\link[gadget3]{g3_to_tmb}}. The model must include \code{\link[gadget3]{g3a_report_detail}}.}

\item{params}{The fitted parameters post-optimisation}

\item{rec.steps}{Which steps should be considered recruitment steps? Vector of int, or NULL for all steps.}

\item{steps}{Which steps to include in the annual output? Vector of int.}
}
\value{
List of tibbles
}
\description{
Build output summary suitable for feeding into gadgetplots
}
