% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/parameter_template_utils.R
\name{g3_init_guess}
\alias{g3_init_guess}
\title{Set initial guesses, automatically honouring any bounded() parameters}
\usage{
g3_init_guess(params, pattern, value = 0, lower = NA, upper = NA, optimise = 1)
}
\arguments{
\item{params}{Parameter template from g3_to_tmb()}

\item{pattern}{Regular expression of parameter names in template to change}

\item{value}{Initial value for all parameters matching pattern}

\item{lower}{Lower bound to set for all parameters matching pattern}

\item{upper}{Upper bound to set for all parameters matching pattern}

\item{optimise}{1 if parameters matching pattern should be optimised, zero otherwise.}
}
\value{
Updated parameter template
}
\description{
Set initial guesses, automatically honouring any bounded() parameters
}
