% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/g3_iterative.R
\name{g3_iterative_final}
\alias{g3_iterative_final}
\title{Updates weights for second stage of iterative re-weighting. Replaced by g3_update_weights}
\usage{
g3_iterative_final(lik_out_list)
}
\arguments{
\item{lik_out_list}{description}
}
\description{
Updates weights for second stage of iterative re-weighting. Replaced by g3_update_weights
}
