% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/g3_iterative.R
\name{g3_iterative_setup}
\alias{g3_iterative_setup}
\title{Initial parameters for iterative re-weighting}
\usage{
g3_iterative_setup(lik_out, grouping = list())
}
\arguments{
\item{lik_out}{A likelihood summary dataframe. The output of g3_lik_out(model, param)}

\item{grouping}{A list describing how to group likelihood components for iterative re-weighting}
}
\value{
A list of G3 parameter dataframes, one for each group of likelihood components
}
\description{
Initial parameters for iterative re-weighting
}
\details{
Initial weights are calculated by taking the inverse SS and then multiplying these by 10000 for the components that are to be optimised
}
