% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/g3_iterative.R
\name{g3_lik_out}
\alias{g3_lik_out}
\title{Likelihood summary for a model run}
\usage{
g3_lik_out(model, param)
}
\arguments{
\item{model}{A G3 model, produced by g3_to_r() or g3_tmb_adfun()}

\item{param}{A G3 parameter dataframe}
}
\value{
A dataframe containing the likelihood component name, degrees of freedom, NLL value, and weight
}
\description{
Likelihood summary for a model run
}
\details{
Runs the model, sums the number of data points (df) and nll's for each component (value), and extracts the corresponding weights from the input parameters (weight)
}
