% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/g3_optim.R
\name{g3_optim}
\alias{g3_optim}
\title{Wrapper for stats::optim that returns the g3_cpp parameter template with optimised values}
\usage{
g3_optim(
  model,
  params,
  use_parscale = TRUE,
  method = "BFGS",
  control = list(),
  print_status = FALSE,
  print_id = "",
  ...
)
}
\arguments{
\item{model}{Either a g3 model of class 'g3_cpp', i.e., model <- gadget3::g3_to_tmb(actions), or an objective function from gadget3::g3_tmb_adfun(model, param)}

\item{params}{The parameter template for a g3_cpp classed model.}

\item{use_parscale}{Should parscale be used g3_tmb_parscale(params), see \code{\link[stats]{optim}}}

\item{method}{The optimisation method, see \code{\link[stats]{optim}}}

\item{control}{List of control options for optim, see \code{\link[stats]{optim}}}

\item{print_status}{Print step comments, does not suppress comments from optim call (see control$trace)}

\item{print_id}{A character string appended to the print statements (useful for when running g3_optim in parallel)}

\item{...}{Further arguments to be passed gadget3::g3_tmb_adfun, see \code{\link[gadget3]{g3_tmb_adfun}}}
}
\value{
A g3_cpp parameter template with optimised values
}
\description{
Wrapper for stats::optim that returns the g3_cpp parameter template with optimised values
}
