% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/g3_retro.R
\name{g3_retro}
\alias{g3_retro}
\title{Analytical retrospective}
\usage{
g3_retro(
  gd,
  outdir = "RETRO",
  model,
  params,
  nyears = 5,
  use_parscale = TRUE,
  method = "BFGS",
  control = list(),
  serial_compile = FALSE,
  mc.cores = parallel::detectCores()
)
}
\arguments{
\item{gd}{Directory to store output}

\item{outdir}{Directory name within gd to store run outputs}

\item{model}{A G3 model, produced by g3_to_r() or g3_tmb_adfun()}

\item{params}{Initial parameters to use with the model, this should be a TMB parameter template i.e. attr(tmb_model, 'parameter_template')}

\item{nyears}{How many years back to go}

\item{use_parscale}{Logical indicating whether optim(control$parscale) should be used}

\item{method}{The optimisation method, see \code{\link[stats]{optim}}}

\item{control}{List of control options for optim, see \code{\link[stats]{optim}}}

\item{serial_compile}{g3_tmb_adfun will be run in serial mode (i.e., not in parallel), potentially helping with memory issues}

\item{mc.cores}{The number of cores to use, defaults to the number available}
}
\value{
List optimised TMB parameter data.frames, the name of each list element corresponds to the number of years peeled in the retrospective analysis.
}
\description{
\code{g3_retro} runs an analytical retrospective model fitting run.
}
