% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/g3_sensitivity.R
\name{g3_sensitivity}
\alias{g3_sensitivity}
\title{Sensitivity analysis}
\usage{
g3_sensitivity(
  gd,
  outdir = "SENS",
  model,
  params,
  par_name,
  par_values,
  use_parscale = TRUE,
  method = "BFGS",
  control = list(),
  serial_compile = FALSE,
  return_input_parameters = FALSE,
  run_optimisations = TRUE,
  mc.cores = parallel::detectCores()
)
}
\arguments{
\item{gd}{Directory to store output}

\item{outdir}{Directory name within gd to store run outputs}

\item{model}{A G3 model, produced by g3_to_tmb() or g3_to_r()}

\item{params}{Initial parameters to use with the model, this should be a TMB parameter template i.e. attr(tmb_model, 'parameter_template')}

\item{par_name}{A glob-like string to match parameter names, see the 'name_spec' argument for \code{\link[gadget3]{g3_init_val}}}

\item{par_values}{A vector of values to substitute into params}

\item{use_parscale}{Logical indicating whether optim(control$parscale) should be used}

\item{method}{The optimisation method, see \code{\link[stats]{optim}}}

\item{control}{List of control options for optim, see \code{\link[stats]{optim}}}

\item{serial_compile}{g3_tmb_adfun will be run in serial mode (i.e., not in parallel), potentially helping with memory issues}

\item{return_input_parameters}{If TRUE, returns a list of modified parameter dataframes (i.e. without optimisations)}

\item{run_optimisations}{If TRUE, will return a list of fits using the modified (non-optimised) parameters}

\item{mc.cores}{The number of cores to use, defaults to the number available}
}
\value{
A list of parameter dataframes (either optimised or not) or a list of fits based on the unoptimised parameters
}
\description{
\code{g3_sensitivity} performs a sensitivity analysis for a single parameter.
}
