% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/parameter_utils.R
\name{g3_stock_name}
\alias{g3_stock_name}
\title{Extract parts from stock name}
\usage{
g3_stock_name(stock, id = "full")
}
\arguments{
\item{stock}{A g3 stock object}

\item{id}{Part to extract, or 'full' to return the stock name}
}
\value{
Part from stock name
}
\description{
Extract parts from stock name
}
