% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/parameter_utils.R
\name{g3_stock_param}
\alias{g3_stock_param}
\title{Generate g3 code to reference a single parameter}
\usage{
g3_stock_param(stock, id = "full", param_name, bound_param = FALSE)
}
\arguments{
\item{stock}{A g3 stock object}

\item{id}{Part of the stock name to use in parameter name}

\item{param_name}{Parameter name to append to the stock name}

\item{bound_param}{Should this parameter be normalised with g3 bounded() ?}
}
\value{
formula to insert parameter into g3 model
}
\description{
Generate g3 code to reference a single parameter
}
