% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/parameter_utils.R
\name{g3_stock_table}
\alias{g3_stock_table}
\title{Generate g3 code to reference a stock parameter, by stock age}
\usage{
g3_stock_table(
  stock,
  id = "full",
  param_name,
  bound_param = FALSE,
  exponentiate = FALSE
)
}
\arguments{
\item{stock}{A g3 stock object}

\item{id}{Part of the stock name to use in parameter name}

\item{param_name}{Parameter name to append to the stock name}

\item{bound_param}{Should this parameter be normalised with g3 bounded() ?}

\item{exponentiate}{Should the parameter be exponentiated}
}
\value{
formula to insert parameter table into g3 model
}
\description{
Generate g3 code to reference a stock parameter, by stock age
}
