% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/g3_iterative.R
\name{g3_update_weights}
\alias{g3_update_weights}
\title{Updates weights for second stage of iterative re-weighting}
\usage{
g3_update_weights(lik_out_list, grouping, cv_floor)
}
\arguments{
\item{lik_out_list}{A list of likelihood summarys. The output of g3_lik_out(model, param) for each group of likelihood components}

\item{grouping}{A list describing how to group likelihood components for iterative re-weighting}

\item{cv_floor}{Minimum value of survey components (adist_surveyindices) as 1/\code{cv_floor}, applied prior to second stage of iterations.}
}
\value{
A list of G3 parameter dataframes, one for each group of likelihood components
}
\description{
Updates weights for second stage of iterative re-weighting
}
\details{
This function updates weights for the second stage of iterative re-weighting. If the cv_floor is > 0, it will be applied here.
}
