% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/g3f_functions.R
\name{g3f_catchdist.fleets}
\alias{g3f_catchdist.fleets}
\title{Summarises the fit of a gadget3 model to distribution data from landings/samples}
\usage{
g3f_catchdist.fleets(reports)
}
\arguments{
\item{reports}{Reported output of a G3 model}
}
\value{
Data frame or NULL
}
\description{
Summarises the fit of a gadget3 model to distribution data from landings/samples
}
