% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/g3f_functions.R
\name{g3f_catchdistribution}
\alias{g3f_catchdistribution}
\title{Summarises the fit of a gadget3 model to catch distributions}
\usage{
g3f_catchdistribution(reports, by_stock = TRUE, by_fleet = TRUE)
}
\arguments{
\item{reports}{Reported output of a G3 model}

\item{by_stock}{Should the stock distribution be compiled?}

\item{by_fleet}{Should the catch distribution per fleet be compiled?}
}
\value{
list containing two data frames: catchdist.fleets and stockdist
}
\description{
Summarises the fit of a gadget3 model to catch distributions
}
