% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/g3f_functions.R
\name{g3f_sidat}
\alias{g3f_sidat}
\title{Summarises the fit of a gadget3 model to survey indices}
\usage{
g3f_sidat(reports)
}
\arguments{
\item{reports}{Reported output of a G3 model}
}
\value{
Data frame or NULL
}
\description{
Summarises the fit of a gadget3 model to survey indices
}
