% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/g3f_functions.R
\name{g3f_stock.recruitment}
\alias{g3f_stock.recruitment}
\title{Summarises the renewal and spawning reports for a gadget 3 model}
\usage{
g3f_stock.recruitment(reports, stock_rec_step = NULL)
}
\arguments{
\item{reports}{Reported output of a G3 model}

\item{stock_rec_step}{Named list where each element contains the step(s) for a specific stock (name of the element). E.g. list(bli_imm = 1, bli_dummy = 2). If NULL or a stock is not included/misspelled, all steps are returned.}
}
\value{
Data frame or NULL
}
\description{
Summarises the renewal and spawning reports for a gadget 3 model
}
