% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/g3f_functions.R
\name{g3f_suitability}
\alias{g3f_suitability}
\title{Extracts the suitabilities from the reports from a gadget3 model}
\usage{
g3f_suitability(reports, stock_names = NULL)
}
\arguments{
\item{reports}{Reported output of a G3 model}

\item{stock_names}{Vector of stock names for which the suits will be retrieved. If NULL, all stocks reported will be included.}
}
\value{
Data frame
}
\description{
Extracts the suitabilities from the reports from a gadget3 model
}
