% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gadget_plots.R
\name{gadget_plots}
\alias{gadget_plots}
\title{Wrapper for plot.gadget.fit that saves all diagnostic graphs to a directory}
\usage{
gadget_plots(fit, path = NULL)
}
\arguments{
\item{fit}{A gadget fit object (can be from gadget2 or gadget3)}

\item{path}{Directory path for saving figures, defaults to file.path(getwd(), 'figs')}
}
\value{
List of results for each optimisation run
}
\description{
Wrapper for plot.gadget.fit that saves all diagnostic graphs to a directory
}
