% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/setup-utils.R
\name{init_abund}
\alias{init_abund}
\title{Generate parameterised g3a_initial_abund}
\usage{
init_abund(
  imm,
  mat,
  comp_id = "species",
  mature = TRUE,
  init_mode = 1,
  bound_param = TRUE
)
}
\arguments{
\item{imm}{A g3 stock object for mature specimens}

\item{comp_id}{Part of stock name to use for parameters, e.g. 'species' will share parameters with both mature/immature}

\item{mature}{Generate actions for mature (TRUE) or immature (FALSE) stock}

\item{init_mode}{One of 0 (initialised at equilibrium), 1 (Initial parameter per age group (across stocks)), 2 (Initial parameter per age group per stock)}

\item{bound_param}{Should parameters be normalised with g3 bounded() ?}
}
\value{
g3a_initial_abund formula for use in g3a_initialconditions_normalparam()
}
\description{
Generate parameterised g3a_initial_abund
}
