% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/setup-utils.R
\name{init_vonb}
\alias{init_vonb}
\title{Generate parameterised version of g3a_initial_vonb()}
\usage{
init_vonb(stock, id = "species", bound_param = TRUE)
}
\arguments{
\item{stock}{A g3 stock object}

\item{id}{Part of the stock name to use in parameter name}

\item{bound_param}{Should this parameter be normalised with g3 bounded() ?}
}
\value{
g3a_initial_vonb formula
}
\description{
Generate parameterised version of g3a_initial_vonb()
}
