% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/g3_jitter.R
\name{jitter_param}
\alias{jitter_param}
\title{Jitters a parameters values}
\usage{
jitter_param(param, jitter_fraction = 0.05)
}
\arguments{
\item{param}{A row from a g3 tmb parameter template}

\item{jitter_fraction}{The fraction of jittering for a value}
}
\value{
Jittered set of parameters
}
\description{
Jitters a parameters values
}
