% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/g3_jitter.R
\name{jitter_params}
\alias{jitter_params}
\title{Creates a jittered g3_tmb parameter template}
\usage{
jitter_params(
  params,
  jitter_fraction = 0.1,
  pattern_to_ignore = NULL,
  within_bounds = TRUE
)
}
\arguments{
\item{params}{A g3 tmb parameter template}

\item{jitter_fraction}{The fraction of jittering for a value}

\item{pattern_to_ignore}{Regular expression of parameters to avoid jittering}

\item{within_bounds}{Logical, if TRUE, jittered values that fall outside parameter bounds (params$lower, params$upper) will be adjusted to fall within the bounds}
}
\value{
Jittered set of parameters
}
\description{
Creates a jittered g3_tmb parameter template
}
