% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/action_utils.R
\name{locate_g3_actions}
\alias{locate_g3_actions}
\title{Find steps relevant to an action type}
\usage{
locate_g3_actions(action_list, type, model_object = NULL)
}
\arguments{
\item{action_list}{List of g3 actions, e.g. list(g3a_time(), ...)}

\item{type}{A string representing the sort of action you're looking for, e.g. "spawning"}

\item{model_object}{The g3_stock object to find an action for}
}
\value{
A data.frame containing pos (position by integer), name & regex pattern for matching steps
}
\description{
Find steps relevant to an action type
}
