% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mohnrho.R
\name{mohnrho}
\alias{mohnrho}
\title{Generate parameterised g3a_initial_abund}
\usage{
mohnrho(fit, retrofit, vars = c("total.biomass", "F", "recruitment"))
}
\arguments{
\item{fit}{A g3_fit object}

\item{retrofit}{A list of g3_fit objects that differ in their terminal year, i.e. the result of a retrospective analysis}

\item{vars}{Which variables would you like to calculate Mohn's Rho for? They must be columns in fit$res.by.year}
}
\value{
A data.frame with mohn's rho for each variable by stock
}
\description{
Generate parameterised g3a_initial_abund
}
