% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/g3_readwrite.R
\name{read.g3.param}
\alias{read.g3.param}
\title{Reads a file from write.g3.file to an object. Now depreciated, using RData files for read/write.}
\usage{
read.g3.param(gd, file.name)
}
\arguments{
\item{gd}{Directory to write the file}

\item{file.name}{Name of the file}
}
\description{
Reads a file from write.g3.file to an object. Now depreciated, using RData files for read/write.
}
