% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/g3_optim.R
\name{run_g3_optim}
\alias{run_g3_optim}
\title{Wrapper for g3_optim. This is an internal function used by the g3_* functions}
\usage{
run_g3_optim(
  model,
  params,
  use_parscale,
  method,
  control,
  serial_compile,
  mc.cores,
  ...
)
}
\arguments{
\item{model}{Either a g3 model of class 'g3_cpp', i.e., model <- gadget3::g3_to_tmb(actions), or an objective function from gadget3::g3_tmb_adfun(model, param)}

\item{params}{A list of parameter templates for a g3_cpp classed model.}

\item{use_parscale}{Should parscale be used g3_tmb_parscale(params), see \code{\link[stats]{optim}}}

\item{method}{The optimisation method, see \code{\link[stats]{optim}}}

\item{control}{List of control options for optim, see \code{\link[stats]{optim}}}

\item{serial_compile}{g3_tmb_adfun will be run in serial mode (i.e., not in parallel), potentially helping with memory issues}

\item{mc.cores}{number of cores used, defaults to the number of available cores}

\item{...}{Further arguments to be passed gadget3::g3_tmb_adfun, see \code{\link[gadget3]{g3_tmb_adfun}}}
}
\value{
A g3_cpp parameter template with optimised values
}
\description{
Wrapper for g3_optim. This is an internal function used by the g3_* functions
}
