% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/parameter_utils.R
\name{stock_species}
\alias{stock_species}
\title{Extract species name from stock}
\usage{
stock_species(stock)
}
\arguments{
\item{stock}{A g3 stock object}
}
\value{
Species name from stock
}
\description{
Extract species name from stock
}
