% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/g3_iterative.R
\name{tabulate_SS}
\alias{tabulate_SS}
\title{Calculates likelihood component scores, including scores normalised by the minimum value}
\usage{
tabulate_SS(lik_out, grouping)
}
\arguments{
\item{lik_out}{Likelihood summary for a model run. Output of g3_lik_out(model, param)}

\item{grouping}{A list describing how to group likelihood components for iterative re-weighting}
}
\value{
A list with a table of likelihood component scores (SS) and a table of normalised likelihood component scores (SS_norm)
}
\description{
Calculates likelihood component scores, including scores normalised by the minimum value
}
