% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/parameter_template_utils.R
\name{value_from_bounds}
\alias{value_from_bounds}
\title{Convert value into it's normalised form for use with bounded()}
\usage{
value_from_bounds(x, lower, upper)
}
\arguments{
\item{x}{Raw value}

\item{lower}{Lower bound for x}

\item{upper}{Upper bound for x}
}
\value{
Normalised value for use with bounded()
}
\description{
Convert value into it's normalised form for use with bounded()
}
