% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/g3_readwrite.R
\name{write.g3.file}
\alias{write.g3.file}
\title{Writes an object to a file, typically just used within gadgetutils g3_ functions}
\usage{
write.g3.file(obj, gd, file_name)
}
\arguments{
\item{obj}{An object to write to file}

\item{gd}{Directory to write the file}

\item{file_name}{Name of the file}
}
\description{
Writes an object to a file, typically just used within gadgetutils g3_ functions
}
