% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/g3_readwrite.R
\name{write.g3.param}
\alias{write.g3.param}
\title{A wrapper for write.g3.file that formats a gadget3 TMB parameter data.frame for writing to file}
\usage{
write.g3.param(params, gd, file_name)
}
\arguments{
\item{params}{A TMB parameter data.frame}

\item{gd}{Directory to write the file}

\item{file_name}{Name of the file}
}
\description{
A wrapper for write.g3.file that formats a gadget3 TMB parameter data.frame for writing to file
}
