% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/add_g3_attributes.R
\name{add_g3_attributes}
\alias{add_g3_attributes}
\title{Add gadget3 attributes to an existing data frame}
\usage{
add_g3_attributes(x, params)
}
\arguments{
\item{x}{A data frame, tibble or data table.}

\item{params}{A list of parameters defining the grouping. See details in \code{mfdb_sample_count}.}
}
\value{
Returns \code{x} with g3 attributes and missing columns as instructed by the \code{params} argument.
}
\description{
Adds gadget3 attributes and missing columns to a a data frame ready to be passed to gadget3. Useful for importing data manually to gadget without the mfdb step.
}
