% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/g3_fit.R
\name{g3_fit}
\alias{g3_fit}
\title{Build output summary suitable for feeding into gadgetplots}
\usage{
g3_fit(model, params, stock_rec_step = NULL, steps = 1, printatstart = 1)
}
\arguments{
\item{model}{A G3 model built using \code{\link[gadget3]{g3_to_r}} or \code{\link[gadget3]{g3_to_tmb}}. The model must include \code{\link[gadget3]{g3a_report_detail}}.}

\item{params}{The fitted parameters post-optimisation}

\item{stock_rec_step}{Named list where each element contains the step(s) for a specific stock (name of the element). E.g. list(bli_imm = 1, bli_dummy = 2). If NULL or a stock is not included/misspelled, all steps are returned.}

\item{steps}{Which steps to include in the annual output? Vector of int.}

\item{printatstart}{Should the stock standard be printed at the start of the timestep (1) or the end (0)}
}
\value{
List of tibbles
}
\description{
Build output summary suitable for feeding into gadgetplots
}
