% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/g3_iterative.R
\name{g3_iterative}
\alias{g3_iterative}
\title{Perform multiple optimisation runs of a model, reweighting with each run}
\usage{
g3_iterative(
  gd,
  wgts = "WGTS",
  model,
  params.in,
  grouping = list(),
  use_parscale = TRUE,
  method = "BFGS",
  control = list()
)
}
\arguments{
\item{gd}{Directory to store output}

\item{wgts}{Directory name within gd to store run outputs}

\item{model}{A G3 model, produced by g3_to_r()}

\item{params.in}{Initial parameters to use with the model}

\item{grouping}{List of component names to optmise together}

\item{use_parscale}{Logical indicating whether optim(control(parscale)) should be used}

\item{method}{The optimisation method, see \code{\link[stats]{optim}}}

\item{control}{List of control options for optim, see \code{\link[stats]{optim}}}
}
\value{
Final set of parameters
}
\description{
Perform multiple optimisation runs of a model, reweighting with each run
}
