% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/g3_jitter.R
\name{g3_jitter}
\alias{g3_jitter}
\title{Jitters a parameters values}
\usage{
g3_jitter(params, jitter_fraction = 0.1, patterns_to_ignore = "_weight")
}
\arguments{
\item{params}{A g3 tmb parameter template}

\item{jitter_fraction}{The fraction of jittering for a value}

\item{patterns_to_ignore}{Regular expression of parameters to avoid jittering}
}
\value{
Jittered set of parameters
}
\description{
Jitters a parameters values
}
