% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/g3_leaveout.R
\name{g3_leaveout}
\alias{g3_leaveout}
\title{Perform a leave-one-out analysis. This involves running a set of optimisations,
each one of which has a component (or set of components) turned off (weight = 0).}
\usage{
g3_leaveout(
  gd,
  outdir = "LOCV",
  model,
  params,
  grouping = list(),
  use_parscale = TRUE,
  method = "BFGS",
  control = list(),
  ncores = parallel::detectCores()
)
}
\arguments{
\item{gd}{Directory to store output}

\item{outdir}{Directory name within gd to store run outputs}

\item{model}{A G3 model, produced by g3_to_tmb() or g3_to_r()}

\item{params}{Initial parameters to use with the model, this should be a TMB parameter template i.e. attr(tmb_model, 'parameter_template')}

\item{grouping}{List of component names to group together. This allows you to leave out multiple components at a time, e.g. length and age-length distributions from a particular fleet}

\item{use_parscale}{Logical indicating whether optim(control$parscale) should be used}

\item{method}{The optimisation method, see \code{\link[stats]{optim}}}

\item{control}{List of control options for optim, see \code{\link[stats]{optim}}}

\item{ncores}{The number of cores to use, defaults to the number available}
}
\value{
A data.frame with the nll for each optimisation
}
\description{
Perform a leave-one-out analysis. This involves running a set of optimisations,
each one of which has a component (or set of components) turned off (weight = 0).
}
