% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/g3_functions.R
\name{g3_retro}
\alias{g3_retro}
\title{Perform multiple optimisation runs of a model, removing a year for each}
\usage{
g3_retro(dir, model, params, num.years = 5)
}
\arguments{
\item{dir}{Directory to store output from RETRO runs}

\item{model}{A G3 model, produced by g3_to_tmb()}

\item{params}{Initial parameters to use with the model}

\item{num.years}{How many years back to go}
}
\value{
List of results for each optimisation run
}
\description{
Perform multiple optimisation runs of a model, removing a year for each
}
