% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/setup-utils.R
\name{init_abund}
\alias{init_abund}
\title{Generate parameterised g3a_initial_abund}
\usage{
init_abund(
  imm,
  mat,
  comp_id = "species",
  mature = TRUE,
  init_mode = 2,
  exp_init = FALSE,
  exp_init_f = FALSE,
  naturalmortality = g3_parameterized("M", by_stock = TRUE)
)
}
\arguments{
\item{imm}{A g3 stock object for mature specimens}

\item{comp_id}{Part of stock name to use for parameters, e.g. 'species' will share parameters with both mature/immature}

\item{mature}{Generate actions for mature (TRUE) or immature (FALSE) stock}

\item{init_mode}{One of 0 (initialised at equilibrium), 1 (Initial parameter per age group (across stocks)), 2 (Initial parameter per age group per stock)}

\item{exp_init}{Logical, whether the initial parameters should be exponentiated}

\item{exp_init_f}{Logical, whether the parameter 'init.f' should be exponentiated}

\item{naturalmortality}{A g3 parameter for natural mortality}
}
\value{
g3a_initial_abund formula for use in g3a_initialconditions_normalparam()
}
\description{
Generate parameterised g3a_initial_abund
}
