% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/setup-utils.R
\name{init_abund}
\alias{init_abund}
\title{Generate parameterised g3a_initial_abund}
\usage{
init_abund(
  imm,
  mat,
  comp_id = "species",
  mature = TRUE,
  init_mode = 2,
  exp_params = c()
)
}
\arguments{
\item{imm}{A g3 stock object for mature specimens}

\item{comp_id}{Part of stock name to use for parameters, e.g. 'species' will share parameters with both mature/immature}

\item{mature}{Generate actions for mature (TRUE) or immature (FALSE) stock}

\item{init_mode}{One of 0 (initialised at equilibrium), 1 (Initial parameter per age group (across stocks)), 2 (Initial parameter per age group per stock)}

\item{exp_params}{Which parameters should be exponentiated? exp_params is a vector of parameter names, possible parameters include: c('init','init.scalar','init.f','m'). Note that is a scalar is exponentiated the annual values will be too, and vice versa.}
}
\value{
g3a_initial_abund formula for use in g3a_initialconditions_normalparam()
}
\description{
Generate parameterised g3a_initial_abund
}
