% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/setup-utils.R
\name{init_sd}
\alias{init_sd}
\title{Generate parameterised initial conditions}
\usage{
init_sd(stock, id, parametric = FALSE, mean_len)
}
\arguments{
\item{stock}{A g3 stock object}

\item{id}{Part of the stock name to use in parameter name}

\item{parametric}{Is the initial conditions stddev parameterised, or a table by age?}

\item{mean_len}{The mean length function (only relevant if parametric = TRUE)}
}
\value{
A formula suitable for g3a_initialconditions_normalparam()
}
\description{
Generate parameterised initial conditions
}
