% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/setup-utils.R
\name{model_actions}
\alias{model_actions}
\title{Generate a standard set of model actions}
\usage{
model_actions(
  imm,
  mat,
  mlgg = 15,
  mature = TRUE,
  comp_id = "species",
  init_mode = 1,
  bound_param = TRUE,
  parametric_sd = FALSE,
  exp_rec = FALSE,
  exp_init = FALSE,
  tv_params = c()
)
}
\arguments{
\item{imm}{A g3 stock object for immature specimens}

\item{mat}{A g3 stock object for mature specimens}

\item{mlgg}{maxlengthgroupgrowth for growth of both mature/immature}

\item{mature}{Generate actions for mature (TRUE) or immature (FALSE) stock}

\item{comp_id}{Part of stock name to use for parameters, e.g. 'species' will share parameters with both mature/immature}

\item{init_mode}{}

\item{bound_param}{Should parameters be normalised with g3 bounded() ?}

\item{parametric_sd}{Is the initial conditions stddev parameterised, or a table by age?}

\item{exp_rec}{Should the recruitment parameters and scalar be exponentiated?}

\item{exp_init}{Should the initial abundance parameters and scalar be exponentiated?}

\item{tv_params}{Which parameters should be time-varying? tv_params is a vector of parameter names, possible time-varying parameters include: 'linf','k','walpha','beta','bbin','recl','rec.sd','mat1','mat2','m'}
}
\value{
A list of g3 actions
}
\description{
Generate a standard set of model actions
}
