% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/setup-utils.R
\name{stock_renewal}
\alias{stock_renewal}
\title{Generate paramerterised stock renewal factor}
\usage{
stock_renewal(stock, id = "species", exponentiate = FALSE)
}
\arguments{
\item{stock}{A g3 stock object}

\item{id}{Part of the stock name to use in parameter name}

\item{exponentiate}{Should the renewal parameters be exponentiated?}
}
\value{
"scalar * renew" formula, parameterised for given stock
}
\description{
Generate paramerterised stock renewal factor
}
